重构: 拆分任务逻辑到独立的文件中
