+++ local_utl/import_feed/CMakeLists.txt
@@ -1,3 +1,3 @@
-add_executable(import_feed main.c)
+add_executable(import_feed main.c test_board.c)
 target_link_libraries(import_feed ythtbbs)
